pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "transformers_knn_adapter"
version = "0.5.0"
description = "Hugging Face Transformers image embedding adapter with a scikit-learn KNN classification head."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "transformers-knn-adapter contributors" }]
dependencies = [
  "transformers",
  "datasets",
  "scikit-learn",
  "numpy",
  "torch",
  "pillow",
  "joblib",
  "tqdm",
  "pytorch-metric-learning>=2.9.0",
  "wandb>=0.25.1",
]

[dependency-groups]
dev = [
  "mypy",
  "ruff",
  "pytest",
  "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/transformers_knn_adapter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
markers = [
  "integration: marks tests that require external model downloads or network access",
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
files = ["src"]
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
disable_error_code = ["override"]
